requirements.txt
asgiref==3.7.2
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
